+++ src/Plugin.php
@@ -53,7 +53,7 @@ class Plugin extends BasePlugin
     /**
      * @inheritdoc
      */
-    public string $schemaVersion = '3.0.0';
+    public string $schemaVersion = '3.0.1';
 
     /**
      * @inheritdoc